requirements.txt
SQLAlchemy>=1.3.8
jsonpickle==1.4
kafka-python>=1.4.6